Renamed load to load_long.
